# Optimize test_relaxed_simd_implies_simd128