--- a/test/sse/hello_sse.out
+++ b/test/sse/hello_sse.out
@@ -1,0 +1,1 @@
+Length of a triangle with sides 3 and 4 is 5.000000.