The test `test_relaxed_simd_implies_simd128` was set to compile test_sse.cpp. But that file is really heavy to compile, making the `test_relaxed_simd_implies_simd128` suite runtime very long:

```
[42%] test_relaxed_simd_implies_simd128 (test_core.strict.test_relaxed_simd_implies_simd128) ... ok (0.64s)
[46%] test_relaxed_simd_implies_simd128 (test_core.minimal0.test_relaxed_simd_implies_simd128) ... ok (0.66s)
[50%] test_relaxed_simd_implies_simd128 (test_core.strict_js.test_relaxed_simd_implies_simd128) ... ok (0.68s)
[53%] test_relaxed_simd_implies_simd128 (test_core.instance.test_relaxed_simd_implies_simd128) ... ok (0.69s)
[57%] test_relaxed_simd_implies_simd128 (test_core.core0.test_relaxed_simd_implies_simd128) ... ok (0.69s)
[61%] test_relaxed_simd_implies_simd128 (test_core.lsan.test_relaxed_simd_implies_simd128) ... ok (0.73s)
[65%] test_relaxed_simd_implies_simd128 (test_core.ubsan.test_relaxed_simd_implies_simd128) ... ok (0.83s)
[69%] test_relaxed_simd_implies_simd128 (test_core.core_2gb.test_relaxed_simd_implies_simd128) ... ok (0.89s)
[73%] test_relaxed_simd_implies_simd128 (test_core.core1.test_relaxed_simd_implies_simd128) ... ok (1.18s)
[76%] test_relaxed_simd_implies_simd128 (test_core.asan.test_relaxed_simd_implies_simd128) ... ok (1.23s)
[80%] test_relaxed_simd_implies_simd128 (test_core.core2.test_relaxed_simd_implies_simd128) ... ok (2.51s)
[84%] test_relaxed_simd_implies_simd128 (test_core.wasmfs.test_relaxed_simd_implies_simd128) ... ok (2.53s)
[88%] test_relaxed_simd_implies_simd128 (test_core.corez.test_relaxed_simd_implies_simd128) ... ok (2.65s)
[92%] test_relaxed_simd_implies_simd128 (test_core.cores.test_relaxed_simd_implies_simd128) ... ok (3.00s)
[96%] test_relaxed_simd_implies_simd128 (test_core.core3.test_relaxed_simd_implies_simd128) ... ok (15.52s)
Total core time: 34.546s. Wallclock time: 15.684s. Parallelization: 2.20x.
```

The test intent is to only verify that passing `-msimd128` implies passing `-msse`, so adjust the test to compile a simple hello world program, resulting in:

```
[42%] test_relaxed_simd_implies_simd128 (test_core.minimal0.test_relaxed_simd_implies_simd128) ... ok (0.57s)
[46%] test_relaxed_simd_implies_simd128 (test_core.strict.test_relaxed_simd_implies_simd128) ... ok (0.57s)
[50%] test_relaxed_simd_implies_simd128 (test_core.core1.test_relaxed_simd_implies_simd128) ... ok (0.61s)
[53%] test_relaxed_simd_implies_simd128 (test_core.strict_js.test_relaxed_simd_implies_simd128) ... ok (0.62s)
[57%] test_relaxed_simd_implies_simd128 (test_core.core0.test_relaxed_simd_implies_simd128) ... ok (0.62s)
[61%] test_relaxed_simd_implies_simd128 (test_core.instance.test_relaxed_simd_implies_simd128) ... ok (0.63s)
[65%] test_relaxed_simd_implies_simd128 (test_core.lsan.test_relaxed_simd_implies_simd128) ... ok (0.64s)
[69%] test_relaxed_simd_implies_simd128 (test_core.ubsan.test_relaxed_simd_implies_simd128) ... ok (0.66s)
[73%] test_relaxed_simd_implies_simd128 (test_core.core_2gb.test_relaxed_simd_implies_simd128) ... ok (0.85s)
[76%] test_relaxed_simd_implies_simd128 (test_core.wasmfs.test_relaxed_simd_implies_simd128) ... ok (1.00s)
[80%] test_relaxed_simd_implies_simd128 (test_core.core2.test_relaxed_simd_implies_simd128) ... ok (0.99s)
[84%] test_relaxed_simd_implies_simd128 (test_core.asan.test_relaxed_simd_implies_simd128) ... ok (1.15s)
[88%] test_relaxed_simd_implies_simd128 (test_core.corez.test_relaxed_simd_implies_simd128) ... ok (1.70s)
[92%] test_relaxed_simd_implies_simd128 (test_core.cores.test_relaxed_simd_implies_simd128) ... ok (1.70s)
[96%] test_relaxed_simd_implies_simd128 (test_core.core3.test_relaxed_simd_implies_simd128) ... ok (1.71s)
Total core time: 14.118s. Wallclock time: 1.874s. Parallelization: 7.53x.
```